chore: add/consolodate components into one file
